Removed inheritance of dict
Only headerArrayObj is still inheriting from dict. According to google inherit from dict is bad as get and [] aceess can become out of sync.
In most cases an OrderedDict is now part of the structure which uses the header names as keys.
Added mtime to the Info structure. is_valid now checks that the time stamp of the file did not change.
Still need test for this.
 
Removed the HeaderName from the HeaderArrayObj. This is now part of the HarFileObj. As the same headerArrayObj might be written to the same file or the same headerArrayObj appears multiple times on the the File having the name in the Array will cause trouble. 

As far as I can see everything is OK and the tests run through.