Make HarFileObj read HeaderData only on demand
For big HAR files this is necessary to avoid memory explosion. To be able to do so, upon every access a validity check is performed whether the file info is up to date. If the file has changed in between, the HAR obj is recreated and a warning is printed. Not sure this is the best way but I could not see a better option